ScenarioTestData.newLineSeparated
