+++ engine/flink/executor/src/test/scala/pl/touk/nussknacker/engine/process/runner/FlinkTestMainSpec.scala
@@ -49,7 +49,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
     val input = SimpleRecord("0", 1, "2", new Date(3), Some(4), 5, "6")
     val input2 = SimpleRecord("0", 11, "2", new Date(3), Some(4), 5, "6")
 
-    val results = runFlinkTest(process, TestData.newLineSeparated("0|1|2|3|4|5|6", "0|11|2|3|4|5|6"))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated("0|1|2|3|4|5|6", "0|11|2|3|4|5|6"))
 
     val nodeResults = results.nodeResults
 
@@ -85,7 +85,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
           GraphBuilder.emptySink("out2", "monitor")
         )
 
-    val results = runFlinkTest(process, TestData.newLineSeparated("0|1|2|3|4|5|6", "0|11|2|3|4|5|6"))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated("0|1|2|3|4|5|6", "0|11|2|3|4|5|6"))
 
     results.nodeResults("splitId1") shouldBe List(nodeResult(0, "input" ->
         SimpleRecord("0", 1, "2", new Date(3), Some(4), 5, "6")),
@@ -108,7 +108,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
     val aggregate2 = SimpleRecordWithPreviousValue(input2, 1, "s")
 
 
-    val results = runFlinkTest(process, TestData.newLineSeparated("0|1|2|3|4|5|6", "0|11|2|3|4|5|6"))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated("0|1|2|3|4|5|6", "0|11|2|3|4|5|6"))
 
     val nodeResults = results.nodeResults
 
@@ -148,7 +148,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
         .source("id", "input")
         .emptySink("out", "monitor")
 
-    val results = runFlinkTest(process, TestData.newLineSeparated("0|1|2|3|4|5|6", "0|11|2|3|4|5|6", "0|11|2|3|4|5|6", "0|11|2|3|4|5|6", "0|11|2|3|4|5|6"))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated("0|1|2|3|4|5|6", "0|11|2|3|4|5|6", "0|11|2|3|4|5|6", "0|11|2|3|4|5|6", "0|11|2|3|4|5|6"))
 
     val nodeResults = results.nodeResults
 
@@ -165,7 +165,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
         .filter("filter", "1 / #input.value1 >= 0")
         .emptySink("out", "monitor")
 
-    val results = runFlinkTest(process, TestData.newLineSeparated("0|1|2|3|4|5|6", "1|0|2|3|4|5|6", "2|2|2|3|4|5|6", "3|4|2|3|4|5|6"))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated("0|1|2|3|4|5|6", "1|0|2|3|4|5|6", "2|2|2|3|4|5|6", "3|4|2|3|4|5|6"))
 
     val nodeResults = results.nodeResults
 
@@ -196,7 +196,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
 
     val exceptionConsumerId = UUID.randomUUID().toString
     val config = RecordingExceptionConsumerProvider.configWithProvider(ConfigFactory.load(), exceptionConsumerId)
-    val results = runFlinkTest(process, TestData.newLineSeparated("0|1|2|3|4|5|6", "1|0|2|3|4|5|6", "2|2|2|3|4|5|6", "3|4|2|3|4|5|6"), config)
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated("0|1|2|3|4|5|6", "1|0|2|3|4|5|6", "2|2|2|3|4|5|6", "3|4|2|3|4|5|6"), config)
 
     val nodeResults = results.nodeResults
 
@@ -216,7 +216,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
         .emptySink("out", "monitor")
 
     val run = Future {
-      runFlinkTest(process, TestData.newLineSeparated("2|2|2|3|4|5|6"))
+      runFlinkTest(process, ScenarioTestData.newLineSeparated("2|2|2|3|4|5|6"))
     }
 
     intercept[JobExecutionException](Await.result(run, 10 seconds))
@@ -262,7 +262,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
       .streaming("proc1")
       .source("id", "genericSourceWithCustomVariables", "elements" -> "{'abc'}")
       .emptySink("out", "valueMonitor", "value" -> "#additionalOne + '|' + #additionalTwo")
-    val testData = TestData.newLineSeparated("abc")
+    val testData = ScenarioTestData.newLineSeparated("abc")
 
     val results = runFlinkTest(process, testData)
 
@@ -280,7 +280,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
         .source("id", "input")
         .emptySink("out", "sinkForInts", "value" -> "15 / {0, 1}[0]")
 
-    val results = runFlinkTest(process, TestData.newLineSeparated("2|2|2|3|4|5|6"))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated("2|2|2|3|4|5|6"))
 
     results.exceptions should have length 1
     results.exceptions.head.nodeId shouldBe Some("out")
@@ -299,7 +299,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
 
     def recordWithSeconds(duration: FiniteDuration) = s"0|0|0|${duration.toMillis}|0|0|0"
 
-    val results = runFlinkTest(process, TestData.newLineSeparated(
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated(
       recordWithSeconds(1 second),
       recordWithSeconds(2 second),
       recordWithSeconds(5 second),
@@ -320,7 +320,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
         .source("id", "typedJsonInput", "type" -> """{"field1": "String", "field2": "java.lang.String"}""")
         .emptySink("out", "valueMonitor", "value" -> "#input.field1 + #input.field2")
 
-    val results = runFlinkTest(process, TestData.newLineSeparated("""{"field1": "abc", "field2": "def"}"""))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated("""{"field1": "abc", "field2": "def"}"""))
 
     results.invocationResults("out").map(_.value) shouldBe List("abcdef")
   }
@@ -336,7 +336,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
         "definition" -> "{'field1', 'field2'}", "toFill" -> "#input.value1.toString()", "count" -> countToPass)
       .emptySink("out", "valueMonitor", "value" ->  "#parsed.size + ' ' + #parsed[0].field2")
 
-    val results = runFlinkTest(process, TestData.newLineSeparated(s"0|$valueToReturn|2|3|4|5|6"))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated(s"0|$valueToReturn|2|3|4|5|6"))
 
     results.invocationResults("out").map(_.value) shouldBe List(s"$countToPass $valueToReturn")
   }
@@ -358,7 +358,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
     val recordTrue = "ala|1|2|3|4|5|6"
     val recordFalse = "bela|1|2|3|4|5|6"
 
-    val results = runFlinkTest(process, TestData.newLineSeparated(recordTrue, recordFalse))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated(recordTrue, recordFalse))
 
     val invocationResults = results.invocationResults
 
@@ -388,7 +388,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
     val recC = "c|1|2|3|4|5|6"
 
 
-    val results = runFlinkTest(process, TestData.newLineSeparated(recA, recB, recC))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated(recA, recB, recC))
 
     //TODO: currently e.g. invocation results will behave strangely in this test, because we duplicate inputs and this results in duplicate context ids...
     results.externalInvocationResults("proc2").map(_.value.asInstanceOf[String]).sorted shouldBe List("a", "b", "c", "c").map(_ + "-collectedDuringServiceInvocation")
@@ -454,7 +454,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
       .customNode("componentUseCaseCustomNode", "componentUseCaseCustomNode", "transformerAddingComponentUseCase")
       .emptySink("out", "valueMonitor", "value" -> "{#componentUseCaseService, #componentUseCaseCustomNode}")
 
-    val results = runFlinkTest(process, TestData.newLineSeparated("0|1|2|3|4|5|6"))
+    val results = runFlinkTest(process, ScenarioTestData.newLineSeparated("0|1|2|3|4|5|6"))
 
     results.invocationResults("out").map(_.value) shouldBe List(List(ComponentUseCase.TestRuntime, ComponentUseCase.TestRuntime).asJava)
   }
